refactor(storage): refactor l0 score calculate and decrease the priority of l0_intra
