+++ src/meta/src/hummock/compaction/picker/base_level_compaction_picker.rs
@@ -58,12 +58,16 @@ impl CompactionPicker for LevelCompactionPicker {
         }
 
         debug_assert!(self.target_level == levels.get_level(self.target_level).level_idx as usize);
-        self.pick_multi_level_to_base(
+        if let Some(ret) = self.pick_multi_level_to_base(
             l0,
             levels.get_level(self.target_level),
             level_handlers,
             stats,
-        )
+        ) {
+            return Some(ret);
+        }
+
+        self.pick_l0_intra(l0, &level_handlers[0], stats)
     }
 }
 
@@ -203,6 +207,127 @@ impl LevelCompactionPicker {
             target_level: self.target_level,
             target_sub_level_id: 0,
         })
+    }
+
+    fn pick_l0_intra(
+        &self,
+        l0: &OverlappingLevel,
+        level_handler: &LevelHandler,
+        stats: &mut LocalPickerStatistic,
+    ) -> Option<CompactionInput> {
+        let overlap_strategy = create_overlap_strategy(self.config.compaction_mode());
+
+        for (idx, level) in l0.sub_levels.iter().enumerate() {
+            if level.level_type() != LevelType::Nonoverlapping
+                || level.total_file_size > self.config.sub_level_max_compaction_bytes
+            {
+                continue;
+            }
+
+            if level_handler.is_level_all_pending_compact(level) {
+                continue;
+            }
+
+            let max_compaction_bytes = std::cmp::min(
+                self.config.max_compaction_bytes,
+                self.config.sub_level_max_compaction_bytes,
+            );
+
+            let tier_sub_level_compact_level_count =
+                self.config.level0_sub_level_compact_level_count as usize;
+            // FIXME(li0k): Just workaround, need to use more reasonable way to limit task size
+            let max_depth = std::cmp::max(
+                (self.config.max_compaction_bytes as f64
+                    / self.config.sub_level_max_compaction_bytes as f64) as usize,
+                tier_sub_level_compact_level_count + 1,
+            );
+            let non_overlap_sub_level_picker = NonOverlapSubLevelPicker::new(
+                0,
+                max_compaction_bytes,
+                1,
+                max_depth,
+                self.config.level0_max_compact_file_number,
+                overlap_strategy.clone(),
+            );
+
+            let l0_select_tables_vec = non_overlap_sub_level_picker
+                .pick_l0_multi_non_overlap_level(&l0.sub_levels[idx..], level_handler);
+
+            if l0_select_tables_vec.is_empty() {
+                continue;
+            }
+
+            let mut skip_by_write_amp = false;
+            let mut level_select_tables: Vec<Vec<SstableInfo>> = vec![];
+            // Limit the number of selection levels for the non-overlapping
+            // sub_level at least level0_sub_level_compact_level_count
+            for (plan_index, (compaction_bytes, all_file_count, level_select_sst)) in
+                l0_select_tables_vec.into_iter().enumerate()
+            {
+                if plan_index == 0
+                    && level_select_sst.len()
+                        < self.config.level0_sub_level_compact_level_count as usize
+                {
+                    // first plan level count smaller than limit
+                    break;
+                }
+
+                let mut max_level_size = 0;
+                for level_select_table in &level_select_sst {
+                    let level_select_size = level_select_table
+                        .iter()
+                        .map(|sst| sst.file_size)
+                        .sum::<u64>();
+
+                    max_level_size = std::cmp::max(max_level_size, level_select_size);
+                }
+
+                // This limitation would keep our write-amplification no more than
+                // ln(max_compaction_bytes/flush_level_bytes) /
+                // ln(self.config.level0_sub_level_compact_level_count/2) Here we only use half
+                // of level0_sub_level_compact_level_count just for convenient.
+                let is_write_amp_large =
+                    max_level_size * self.config.level0_sub_level_compact_level_count as u64 / 2
+                        >= compaction_bytes;
+
+                if is_write_amp_large
+                    && level_select_sst.len() < tier_sub_level_compact_level_count
+                    && all_file_count < self.config.level0_max_compact_file_number as usize
+                {
+                    skip_by_write_amp = true;
+                    continue;
+                }
+
+                level_select_tables = level_select_sst;
+                break;
+            }
+
+            if level_select_tables.is_empty() {
+                if skip_by_write_amp {
+                    stats.skip_by_write_amp_limit += 1;
+                }
+                continue;
+            }
+
+            let mut select_level_inputs = Vec::with_capacity(level_select_tables.len());
+            for level_select_sst in level_select_tables {
+                select_level_inputs.push(InputLevel {
+                    level_idx: 0,
+                    level_type: LevelType::Nonoverlapping as i32,
+                    table_infos: level_select_sst,
+                });
+            }
+
+            select_level_inputs.reverse();
+
+            return Some(CompactionInput {
+                input_levels: select_level_inputs,
+                target_level: 0,
+                target_sub_level_id: level.sub_level_id,
+            });
+        }
+
+        None
     }
 }
 
@@ -649,9 +774,11 @@ pub mod tests {
         // Only include sub-level 0 results will violate MAX_WRITE_AMPLIFICATION.
         // But stopped by pending sub-level when trying to include more sub-levels.
         let mut picker = LevelCompactionPicker::new(1, config.clone());
-        assert!(picker
+        let ret = picker
             .pick_compaction(&levels, &levels_handler, &mut local_stats)
-            .is_none());
+            .unwrap();
+        // when l1 pending, try to trigger l0_intra
+        assert!(ret.input_levels.last().unwrap().level_idx == 0);
 
         // Free the pending sub-level.
         for pending_task_id in &levels_handler[0].pending_tasks_ids() {
@@ -709,5 +836,187 @@ pub mod tests {
         assert_eq!(5, ret.input_levels[0].table_infos[0].sst_id);
         assert_eq!(1, ret.input_levels[1].level_idx);
         assert!(ret.input_levels[1].table_infos.is_empty());
+    }
+
+    #[test]
+    fn test_pick_l0_intra() {
+        {
+            let l0 = generate_l0_nonoverlapping_multi_sublevels(vec![
+                vec![
+                    generate_table(6, 1, 50, 99, 1),
+                    generate_table(1, 1, 100, 200, 1),
+                    generate_table(2, 1, 250, 300, 1),
+                ],
+                vec![
+                    generate_table(3, 1, 10, 90, 1),
+                    generate_table(6, 1, 100, 110, 1),
+                ],
+                vec![
+                    generate_table(4, 1, 50, 99, 1),
+                    generate_table(5, 1, 100, 200, 1),
+                ],
+            ]);
+            let levels = Levels {
+                l0: Some(l0),
+                levels: vec![generate_level(1, vec![generate_table(100, 1, 0, 1000, 1)])],
+                member_table_ids: vec![1],
+                ..Default::default()
+            };
+            let mut levels_handler = vec![LevelHandler::new(0), LevelHandler::new(1)];
+            levels_handler[1].add_pending_task(100, 1, levels.levels[0].get_table_infos());
+            let config = Arc::new(
+                CompactionConfigBuilder::new()
+                    .level0_sub_level_compact_level_count(1)
+                    .level0_overlapping_sub_level_compact_level_count(4)
+                    .build(),
+            );
+            let mut picker = LevelCompactionPicker::new(1, config);
+            let mut local_stats = LocalPickerStatistic::default();
+            let ret = picker
+                .pick_compaction(&levels, &levels_handler, &mut local_stats)
+                .unwrap();
+            ret.add_pending_task(1, &mut levels_handler);
+            assert_eq!(
+                ret.input_levels
+                    .iter()
+                    .map(|i| i.table_infos.len())
+                    .sum::<usize>(),
+                3
+            );
+        }
+
+        {
+            // Suppose keyguard [100, 200] [300, 400]
+            // will pick sst [1, 3, 4]
+            let l0 = generate_l0_nonoverlapping_multi_sublevels(vec![
+                vec![
+                    generate_table(1, 1, 100, 200, 1),
+                    generate_table(2, 1, 300, 400, 1),
+                ],
+                vec![
+                    generate_table(3, 1, 100, 200, 1),
+                    generate_table(6, 1, 300, 500, 1),
+                ],
+                vec![
+                    generate_table(4, 1, 100, 200, 1),
+                    generate_table(5, 1, 300, 400, 1),
+                ],
+            ]);
+            let levels = Levels {
+                l0: Some(l0),
+                levels: vec![generate_level(1, vec![generate_table(100, 1, 0, 1000, 1)])],
+                member_table_ids: vec![1],
+                ..Default::default()
+            };
+            let mut levels_handler = vec![LevelHandler::new(0), LevelHandler::new(1)];
+            levels_handler[1].add_pending_task(100, 1, levels.levels[0].get_table_infos());
+            let config = Arc::new(
+                CompactionConfigBuilder::new()
+                    .level0_sub_level_compact_level_count(1)
+                    .build(),
+            );
+            let mut picker = LevelCompactionPicker::new(1, config);
+            let mut local_stats = LocalPickerStatistic::default();
+            let ret = picker
+                .pick_compaction(&levels, &levels_handler, &mut local_stats)
+                .unwrap();
+            ret.add_pending_task(1, &mut levels_handler);
+            assert_eq!(
+                ret.input_levels
+                    .iter()
+                    .map(|i| i.table_infos.len())
+                    .sum::<usize>(),
+                3
+            );
+
+            assert_eq!(4, ret.input_levels[0].table_infos[0].get_sst_id());
+            assert_eq!(3, ret.input_levels[1].table_infos[0].get_sst_id());
+            assert_eq!(1, ret.input_levels[2].table_infos[0].get_sst_id());
+
+            // will pick sst [2, 6, 5]
+            let ret2 = picker
+                .pick_compaction(&levels, &levels_handler, &mut local_stats)
+                .unwrap();
+
+            assert_eq!(
+                ret2.input_levels
+                    .iter()
+                    .map(|i| i.table_infos.len())
+                    .sum::<usize>(),
+                3
+            );
+
+            assert_eq!(5, ret2.input_levels[0].table_infos[0].get_sst_id());
+            assert_eq!(6, ret2.input_levels[1].table_infos[0].get_sst_id());
+            assert_eq!(2, ret2.input_levels[2].table_infos[0].get_sst_id());
+        }
+
+        {
+            let l0 = generate_l0_nonoverlapping_multi_sublevels(vec![
+                vec![
+                    generate_table(1, 1, 100, 149, 1),
+                    generate_table(6, 1, 150, 199, 1),
+                    generate_table(7, 1, 200, 250, 1),
+                    generate_table(2, 1, 300, 400, 1),
+                ],
+                vec![
+                    generate_table(3, 1, 100, 149, 1),
+                    generate_table(8, 1, 150, 199, 1),
+                    generate_table(9, 1, 200, 250, 1),
+                    generate_table(10, 1, 300, 400, 1),
+                ],
+                vec![
+                    generate_table(4, 1, 100, 199, 1),
+                    generate_table(11, 1, 200, 250, 1),
+                    generate_table(5, 1, 300, 350, 1),
+                ],
+            ]);
+            let levels = Levels {
+                l0: Some(l0),
+                levels: vec![generate_level(1, vec![generate_table(100, 1, 0, 1000, 1)])],
+                member_table_ids: vec![1],
+                ..Default::default()
+            };
+            let mut levels_handler = vec![LevelHandler::new(0), LevelHandler::new(1)];
+            levels_handler[1].add_pending_task(100, 1, levels.levels[0].get_table_infos());
+            let config = Arc::new(
+                CompactionConfigBuilder::new()
+                    .level0_sub_level_compact_level_count(1)
+                    .build(),
+            );
+            let mut picker = LevelCompactionPicker::new(1, config);
+            let mut local_stats = LocalPickerStatistic::default();
+            let ret = picker
+                .pick_compaction(&levels, &levels_handler, &mut local_stats)
+                .unwrap();
+            ret.add_pending_task(1, &mut levels_handler);
+            assert_eq!(
+                ret.input_levels
+                    .iter()
+                    .map(|i| i.table_infos.len())
+                    .sum::<usize>(),
+                3
+            );
+
+            assert_eq!(11, ret.input_levels[0].table_infos[0].get_sst_id());
+            assert_eq!(9, ret.input_levels[1].table_infos[0].get_sst_id());
+            assert_eq!(7, ret.input_levels[2].table_infos[0].get_sst_id());
+
+            let ret2 = picker
+                .pick_compaction(&levels, &levels_handler, &mut local_stats)
+                .unwrap();
+
+            assert_eq!(
+                ret2.input_levels
+                    .iter()
+                    .map(|i| i.table_infos.len())
+                    .sum::<usize>(),
+                3
+            );
+
+            assert_eq!(5, ret2.input_levels[0].table_infos[0].get_sst_id());
+            assert_eq!(10, ret2.input_levels[1].table_infos[0].get_sst_id());
+            assert_eq!(2, ret2.input_levels[2].table_infos[0].get_sst_id());
+        }
     }
 }
