+++ src/meta/src/hummock/compaction/picker/tier_compaction_picker.rs
@@ -15,13 +15,10 @@
 use std::sync::Arc;
 
 use risingwave_pb::hummock::hummock_version::Levels;
-use risingwave_pb::hummock::{
-    CompactionConfig, InputLevel, LevelType, OverlappingLevel, SstableInfo,
-};
+use risingwave_pb::hummock::{CompactionConfig, InputLevel, LevelType, OverlappingLevel};
 
 use super::{CompactionInput, CompactionPicker, LocalPickerStatistic, MinOverlappingPicker};
 use crate::hummock::compaction::overlap_strategy::OverlapStrategy;
-use crate::hummock::compaction::picker::min_overlap_compaction_picker::NonOverlapSubLevelPicker;
 use crate::hummock::level_handler::LevelHandler;
 
 pub struct TierCompactionPicker {
@@ -38,125 +35,6 @@ impl TierCompactionPicker {
             config,
             overlap_strategy,
         }
-    }
-
-    fn pick_multi_level(
-        &self,
-        l0: &OverlappingLevel,
-        level_handler: &LevelHandler,
-        stats: &mut LocalPickerStatistic,
-    ) -> Option<CompactionInput> {
-        for (idx, level) in l0.sub_levels.iter().enumerate() {
-            if level.level_type() != LevelType::Nonoverlapping
-                || level.total_file_size > self.config.sub_level_max_compaction_bytes
-            {
-                continue;
-            }
-
-            if level_handler.is_level_all_pending_compact(level) {
-                continue;
-            }
-
-            let max_compaction_bytes = std::cmp::min(
-                self.config.max_compaction_bytes,
-                self.config.sub_level_max_compaction_bytes,
-            );
-
-            let tier_sub_level_compact_level_count =
-                self.config.level0_sub_level_compact_level_count as usize;
-            // FIXME(li0k): Just workaround, need to use more reasonable way to limit task size
-            let max_depth = std::cmp::max(
-                (self.config.max_compaction_bytes as f64
-                    / self.config.sub_level_max_compaction_bytes as f64) as usize,
-                tier_sub_level_compact_level_count + 1,
-            );
-            let non_overlap_sub_level_picker = NonOverlapSubLevelPicker::new(
-                0,
-                max_compaction_bytes,
-                1,
-                max_depth,
-                self.config.level0_max_compact_file_number,
-                self.overlap_strategy.clone(),
-            );
-
-            let l0_select_tables_vec = non_overlap_sub_level_picker
-                .pick_l0_multi_non_overlap_level(&l0.sub_levels[idx..], level_handler);
-
-            if l0_select_tables_vec.is_empty() {
-                continue;
-            }
-
-            let mut skip_by_write_amp = false;
-            let mut level_select_tables: Vec<Vec<SstableInfo>> = vec![];
-            // Limit the number of selection levels for the non-overlapping
-            // sub_level at least level0_sub_level_compact_level_count
-            for (plan_index, (compaction_bytes, all_file_count, level_select_sst)) in
-                l0_select_tables_vec.into_iter().enumerate()
-            {
-                if plan_index == 0
-                    && level_select_sst.len()
-                        < self.config.level0_sub_level_compact_level_count as usize
-                {
-                    // first plan level count smaller than limit
-                    break;
-                }
-
-                let mut max_level_size = 0;
-                for level_select_table in &level_select_sst {
-                    let level_select_size = level_select_table
-                        .iter()
-                        .map(|sst| sst.file_size)
-                        .sum::<u64>();
-
-                    max_level_size = std::cmp::max(max_level_size, level_select_size);
-                }
-
-                // This limitation would keep our write-amplification no more than
-                // ln(max_compaction_bytes/flush_level_bytes) /
-                // ln(self.config.level0_sub_level_compact_level_count/2) Here we only use half
-                // of level0_sub_level_compact_level_count just for convenient.
-                let is_write_amp_large =
-                    max_level_size * self.config.level0_sub_level_compact_level_count as u64 / 2
-                        >= compaction_bytes;
-
-                if is_write_amp_large
-                    && level_select_sst.len() < tier_sub_level_compact_level_count
-                    && all_file_count < self.config.level0_max_compact_file_number as usize
-                {
-                    skip_by_write_amp = true;
-                    continue;
-                }
-
-                level_select_tables = level_select_sst;
-                break;
-            }
-
-            if level_select_tables.is_empty() {
-                if skip_by_write_amp {
-                    stats.skip_by_write_amp_limit += 1;
-                }
-                continue;
-            }
-
-            let mut select_level_inputs = Vec::with_capacity(level_select_tables.len());
-            for level_select_sst in level_select_tables {
-                select_level_inputs.push(InputLevel {
-                    level_idx: 0,
-                    level_type: LevelType::Nonoverlapping as i32,
-                    table_infos: level_select_sst,
-                });
-            }
-
-            select_level_inputs.reverse();
-
-            return Some(CompactionInput {
-                input_levels: select_level_inputs,
-                target_level: 0,
-                target_sub_level_id: level.sub_level_id,
-            });
-        }
-
-        None
     }
 
     fn pick_trivial_move_file(
@@ -343,11 +221,7 @@ impl CompactionPicker for TierCompactionPicker {
             return Some(ret);
         }
 
-        if let Some(ret) = self.pick_trivial_move_file(l0, level_handlers) {
-            return Some(ret);
-        }
-
-        self.pick_multi_level(l0, &level_handlers[0], stats)
+        self.pick_trivial_move_file(l0, level_handlers)
     }
 }
 
@@ -361,8 +235,7 @@ pub mod tests {
 
     use crate::hummock::compaction::compaction_config::CompactionConfigBuilder;
     use crate::hummock::compaction::level_selector::tests::{
-        generate_l0_nonoverlapping_multi_sublevels, generate_l0_overlapping_sublevels,
-        generate_table, push_table_level0_overlapping,
+        generate_l0_overlapping_sublevels, generate_table, push_table_level0_overlapping,
     };
     use crate::hummock::compaction::overlap_strategy::RangeOverlapStrategy;
     use crate::hummock::compaction::picker::{
@@ -503,188 +376,6 @@ pub mod tests {
         assert!(picker
             .pick_compaction(&empty_level, &levels_handler, &mut local_stats)
             .is_none());
-    }
-
-    #[test]
-    fn test_pick_multi_level_basic() {
-        {
-            let l0 = generate_l0_nonoverlapping_multi_sublevels(vec![
-                vec![
-                    generate_table(6, 1, 50, 99, 1),
-                    generate_table(1, 1, 100, 200, 1),
-                    generate_table(2, 1, 250, 300, 1),
-                ],
-                vec![
-                    generate_table(3, 1, 10, 90, 1),
-                    generate_table(6, 1, 100, 110, 1),
-                ],
-                vec![
-                    generate_table(4, 1, 50, 99, 1),
-                    generate_table(5, 1, 100, 200, 1),
-                ],
-            ]);
-            let levels = Levels {
-                l0: Some(l0),
-                levels: vec![],
-                member_table_ids: vec![1],
-                ..Default::default()
-            };
-            let mut levels_handler = vec![LevelHandler::new(0)];
-            let config = Arc::new(
-                CompactionConfigBuilder::new()
-                    .level0_sub_level_compact_level_count(1)
-                    .level0_overlapping_sub_level_compact_level_count(4)
-                    .build(),
-            );
-            let mut picker =
-                TierCompactionPicker::new(config, Arc::new(RangeOverlapStrategy::default()));
-            let mut local_stats = LocalPickerStatistic::default();
-            let ret = picker
-                .pick_compaction(&levels, &levels_handler, &mut local_stats)
-                .unwrap();
-            ret.add_pending_task(1, &mut levels_handler);
-            assert_eq!(
-                ret.input_levels
-                    .iter()
-                    .map(|i| i.table_infos.len())
-                    .sum::<usize>(),
-                3
-            );
-        }
-
-        {
-            // Suppose keyguard [100, 200] [300, 400]
-            // will pick sst [1, 3, 4]
-            let l0 = generate_l0_nonoverlapping_multi_sublevels(vec![
-                vec![
-                    generate_table(1, 1, 100, 200, 1),
-                    generate_table(2, 1, 300, 400, 1),
-                ],
-                vec![
-                    generate_table(3, 1, 100, 200, 1),
-                    generate_table(6, 1, 300, 500, 1),
-                ],
-                vec![
-                    generate_table(4, 1, 100, 200, 1),
-                    generate_table(5, 1, 300, 400, 1),
-                ],
-            ]);
-            let levels = Levels {
-                l0: Some(l0),
-                levels: vec![],
-                member_table_ids: vec![1],
-                ..Default::default()
-            };
-            let mut levels_handler = vec![LevelHandler::new(0)];
-            let config = Arc::new(
-                CompactionConfigBuilder::new()
-                    .level0_sub_level_compact_level_count(1)
-                    .build(),
-            );
-            let mut picker =
-                TierCompactionPicker::new(config, Arc::new(RangeOverlapStrategy::default()));
-            let mut local_stats = LocalPickerStatistic::default();
-            let ret = picker
-                .pick_compaction(&levels, &levels_handler, &mut local_stats)
-                .unwrap();
-            ret.add_pending_task(1, &mut levels_handler);
-            assert_eq!(
-                ret.input_levels
-                    .iter()
-                    .map(|i| i.table_infos.len())
-                    .sum::<usize>(),
-                3
-            );
-
-            assert_eq!(4, ret.input_levels[0].table_infos[0].get_sst_id());
-            assert_eq!(3, ret.input_levels[1].table_infos[0].get_sst_id());
-            assert_eq!(1, ret.input_levels[2].table_infos[0].get_sst_id());
-
-            // will pick sst [2, 6, 5]
-            let ret2 = picker
-                .pick_compaction(&levels, &levels_handler, &mut local_stats)
-                .unwrap();
-
-            assert_eq!(
-                ret2.input_levels
-                    .iter()
-                    .map(|i| i.table_infos.len())
-                    .sum::<usize>(),
-                3
-            );
-
-            assert_eq!(5, ret2.input_levels[0].table_infos[0].get_sst_id());
-            assert_eq!(6, ret2.input_levels[1].table_infos[0].get_sst_id());
-            assert_eq!(2, ret2.input_levels[2].table_infos[0].get_sst_id());
-        }
-
-        {
-            let l0 = generate_l0_nonoverlapping_multi_sublevels(vec![
-                vec![
-                    generate_table(1, 1, 100, 149, 1),
-                    generate_table(6, 1, 150, 199, 1),
-                    generate_table(7, 1, 200, 250, 1),
-                    generate_table(2, 1, 300, 400, 1),
-                ],
-                vec![
-                    generate_table(3, 1, 100, 149, 1),
-                    generate_table(8, 1, 150, 199, 1),
-                    generate_table(9, 1, 200, 250, 1),
-                    generate_table(10, 1, 300, 400, 1),
-                ],
-                vec![
-                    generate_table(4, 1, 100, 199, 1),
-                    generate_table(11, 1, 200, 250, 1),
-                    generate_table(5, 1, 300, 350, 1),
-                ],
-            ]);
-            let levels = Levels {
-                l0: Some(l0),
-                levels: vec![],
-                member_table_ids: vec![1],
-                ..Default::default()
-            };
-            let mut levels_handler = vec![LevelHandler::new(0)];
-            let config = Arc::new(
-                CompactionConfigBuilder::new()
-                    .level0_sub_level_compact_level_count(1)
-                    .build(),
-            );
-            let mut picker =
-                TierCompactionPicker::new(config, Arc::new(RangeOverlapStrategy::default()));
-            let mut local_stats = LocalPickerStatistic::default();
-            let ret = picker
-                .pick_compaction(&levels, &levels_handler, &mut local_stats)
-                .unwrap();
-            ret.add_pending_task(1, &mut levels_handler);
-            assert_eq!(
-                ret.input_levels
-                    .iter()
-                    .map(|i| i.table_infos.len())
-                    .sum::<usize>(),
-                3
-            );
-
-            assert_eq!(11, ret.input_levels[0].table_infos[0].get_sst_id());
-            assert_eq!(9, ret.input_levels[1].table_infos[0].get_sst_id());
-            assert_eq!(7, ret.input_levels[2].table_infos[0].get_sst_id());
-
-            let ret2 = picker
-                .pick_compaction(&levels, &levels_handler, &mut local_stats)
-                .unwrap();
-
-            assert_eq!(
-                ret2.input_levels
-                    .iter()
-                    .map(|i| i.table_infos.len())
-                    .sum::<usize>(),
-                3
-            );
-
-            assert_eq!(5, ret2.input_levels[0].table_infos[0].get_sst_id());
-            assert_eq!(10, ret2.input_levels[1].table_infos[0].get_sst_id());
-            assert_eq!(2, ret2.input_levels[2].table_infos[0].get_sst_id());
-        }
     }
 
     #[test]
